+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *Revisions*
 
 * Improved prompt to Wolf-hound on how to choose an allegiance. ([#90](https://github.com/hjwylde/werewolf/issues/90))
+* Changed Scapegoat's balance to 0. ([#91](https://github.com/hjwylde/werewolf/issues/91))
 
 #### v0.4.7.0